backend/requirements.txt
# HelloAgents框架
hello-agents[protocols]>=0.2.4,<=0.2.9

# FastAPI和相关依赖
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP客户端
httpx>=0.27.0
aiohttp>=3.10.0

# 环境变量管理
python-dotenv>=1.0.0

# CORS支持
python-multipart>=0.0.9

# 日志
loguru>=0.7.0

# MCP相关
fastmcp>=2.0.0
uv>=0.8.0

# 数据库与认证
sqlalchemy>=2.0.36
alembic>=1.13.3
psycopg[binary]>=3.2.3
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.2.0

# 其他工具
python-dateutil>=2.8.2
huggingface_hub>=0.25.0
